travis-ci + tox
