+++ tests/bin/mayapy2018
@@ -1,0 +1,2 @@
+#!/usr/bin/env bash
+docker exec maya mayapy
