+++ tox.ini
@@ -4,14 +4,35 @@
 # and then run "tox" from this directory.
 
 [tox]
-skipsdist=True
+envlist = py27, py36
+skipsdist = True
 
-[testenv:cov]
-usedevelop = true
+[testenv]
 deps =
-    pymongo
     pytest-cov
     pytest-bdd
-    pytest
+    pytest-env-info
+    pymongo
+    pyqt5==5.9
+passenv =
+	PYTHONPATH
+	PATH
+	TEMP
+	TMP
+	TMPDIR
+	LOGNAME
+	USER
+	LNAME
+	USERNAME
+	PATHEXT
+	SYSTEMDRIVE
+	SYSTEMROOT
+	AVALON_MONGO
+	PYBLISH_BASE
+	PYBLISH_QML
+	AVALON_CORE
+	AVALON_LAUNCHER
+	AVALON_SETUP
+	CONFIG_ROOT
 commands =
     pytest --cov-report term-missing --cov reveries --disable-warnings
